requirements.txt
fastapi
uvicorn[standard]
gunicorn
httpx
kafka-python
python-dotenv
pydantic
pydantic-settings
